Time: 5 ms (33.70%), Space: 6.6 MB (91.95%) - LeetHub

+++ 377-combination-sum-iv/377-combination-sum-iv.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int combinationSum4(vector<int>& nums, int t) {
-         vector<unsigned int> dp(t + 1, 0);
+         vector<unsigned> dp(t + 1, 0);
          dp[0] = 1;
     
          for (int i = 1; i <= t; i++){
@@ -11,6 +11,6 @@ public:
              }
          }
     
-    return dp[t];
+    return (int)dp[t];
     }
 };
